requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
Django==1.11
djangorestframework==3.6.2
pytz==2017.2
